Chore(Refactoring): Refactoring routes code [Finishes #168377676]
