+++ server/tests/index.test.js
@@ -1,6 +1,6 @@
 import chai, { expect } from 'chai';
 import chaiHttp from 'chai-http';
-import app from '../../index';
+import app from '../../server';
 
 
 chai.use(chaiHttp);
